feat: Add conservative regridding and fix stri_nearnd (#11)
* feat: Add conservative regridding and fix stri_nearnd

This commit introduces a first-order conservative regridding feature.

The main changes are:
- A new C function `ssrf_conservative_regrid` that uses sub-grid sampling to perform conservative regridding.
- A Python wrapper `regrid_conservative` in the `SphericalMesh` class that calls the C function and renormalizes the output to enforce global conservation.
- A fix for the `stri_nearnd` function, which was previously a stub and is now a fully functional nearest-neighbor search.

This also includes a new test for the regridding feature and updates to the Python wrapper to improve performance by caching `argtypes` and `float32` versions of the coordinate arrays.

* fix: Format code with ruff

This commit fixes formatting issues in `renka/renka.py` and `tests/test_renka.py` identified during code review.

* feat: Add benchmark tests for regrid_conservative

This commit adds benchmark tests for the `regrid_conservative` function to evaluate its performance with varying input sizes.

The new file `tests/test_benchmark.py` contains two tests:
- `test_benchmark_regrid_conservative_scaling_source`: Benchmarks the function with a varying number of source points.
- `test_benchmark_regrid_conservative_scaling_grid`: Benchmarks the function with a varying destination grid size.

* feat: Add benchmark tests for interpolate method

This commit adds benchmark tests for the `interpolate` method to evaluate its performance with varying input sizes.

The new tests in `tests/test_benchmark.py` are:
- `test_benchmark_interpolate_scaling_source`: Benchmarks the function with a varying number of source points.
- `test_benchmark_interpolate_scaling_grid`: Benchmarks the function with a varying destination grid size.

* fix: Format benchmark tests with ruff

This commit fixes formatting issues in `tests/test_benchmark.py` identified during code review.

---------

+++ src/stripack.h
@@ -27,7 +27,7 @@ int stri_jrand(int n, int* ix, int* iy, int* iz);
 int stri_left(double x1, double y1, double z1, double x2, double y2, double z2, double x0, double y0, double z0);
 int stri_lstptr(int lpl, int nb, int* list, int* lptr);
 int stri_nbcnt(int lpl, int* lptr);
-int stri_nearnd(double* p, int ist, int n, double* x, double* y, double* z, int* list, int* lptr, int* lend, double* al);
+int stri_nearnd(float* p, int ist, int n, float* x, float* y, float* z, int* list, int* lptr, int* lend, float* al);
 void stri_optim(double* x, double* y, double* z, int na, int* list, int* lptr, int* lend, int* nit, int* iwk, int* ier);
 double stri_store(double x);
 void stri_swap(int in1, int in2, int io1, int io2, int* list, int* lptr, int* lend, int* lp21);
